Rename auto xxx to xxx lab
