+++ PowerPointLabs/PowerPointLabs/Ribbon1.xml
@@ -20,7 +20,7 @@
       <tab id="PowerPointLabsAddIns"
            getLabel="GetPowerPointLabsAddInsTabLabel">
         <group id="AutoMotionGroup"
-               getLabel="GetAutoAnimateGroupLabel"
+               getLabel="GetAnimationLabGroupLabel"
                visible ="true">
           <button id="AddAnimationButton"
                   getLabel="GetAddAnimationButtonLabel"
@@ -37,12 +37,12 @@
                    getEnabled="OnGetEnabledAddInSlide"
                    onAction="AddInSlideAnimationButtonClick"/>
           <dialogBoxLauncher>
-            <button id ="AutoAnimateDialogBox"
-                    onAction="AutoAnimateDialogButtonPressed"/>
+            <button id ="AnimationLabDialogBox"
+                    onAction="AnimationLabDialogButtonPressed"/>
           </dialogBoxLauncher>
         </group>
-        <group id="AutoZoomGroup"
-               getLabel="GetAutoZoomGroupLabel"
+        <group id="ZoomLabGroup"
+               getLabel="GetZoomLabGroupLabel"
                visible ="true">
           <button id="AddZoomInButton"
                  getLabel ="GetAddZoomInButtonLabel"
@@ -66,12 +66,12 @@
                  getEnabled="OnGetEnabledZoomButton"
                  onAction="ZoomBtnClick"/>
           <dialogBoxLauncher>
-            <button id ="AutoZoomDialogBox"
-                    onAction="AutoZoomDialogButtonPressed"/>
+            <button id ="ZoomLabDialogBox"
+                    onAction="ZoomLabDialogButtonPressed"/>
           </dialogBoxLauncher>
         </group>
-        <group id="AutoCropGroup"
-               getLabel="GetAutoCropGroupLabel"
+        <group id="CropLabGroup"
+               getLabel="GetCropLabGroupLabel"
                visible="true">
           <button id="MoveCropShapeButton"
                   getLabel ="GetMoveCropShapeButtonLabel"
@@ -111,8 +111,8 @@
                   getImage="GetRemoveAudioImage"
                   size="large" />
           <dialogBoxLauncher>
-            <button id="AutoNarrateDialogBox"
-                    onAction="AutoNarrateDialogButtonPressed"/>
+            <button id="NarrationsLabDialogBox"
+                    onAction="NarrationsLabDialogButtonPressed"/>
           </dialogBoxLauncher>
         </group>
         <group id="embedCaptions"
